remove http_request test, restructure
